pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "open-workspace-builder"
version = "0.4.0"
description = "Scaffold, maintain, and secure AI coding workspaces"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{name = "originalrgsec"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# litellm upper bound: versions 1.82.7 and 1.82.8 were compromised in a supply
# chain attack on 2026-03-24 (Snyk SNYK-PYTHON-LITELLM-15762713). Do not remove
# the upper bound until LiteLLM publishes a post-incident report confirming CI
# hardening.
llm = [
    "litellm>=1.40.0,<=1.82.6",
]
security = [
    "pyyaml>=6.0",
    "litellm>=1.40.0,<=1.82.6",
]
keyring = [
    "keyring>=24.0",
]
age = [
    "pyrage>=1.0",
]
secrets = [
    "keyring>=24.0",
    "pyrage>=1.0",
]
audit = [
    "pip-audit>=2.7.0",
]
sast = [
    "semgrep>=1.50.0",
]
mcp = [
    "mcp>=1.0.0",
]
sheets = [
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
]
xlsx = [
    "xlsxwriter>=3.0",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "build",
    "pip-audit>=2.7.0",
    "semgrep>=1.50.0",
]

[project.urls]
Homepage = "https://github.com/originalrgsec/open-workspace-builder"
Repository = "https://github.com/originalrgsec/open-workspace-builder"
Issues = "https://github.com/originalrgsec/open-workspace-builder/issues"

[project.scripts]
owb = "open_workspace_builder.cli:owb"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
open_workspace_builder = ["py.typed", "security/data/*.yaml", "data/registry/**/*.yaml", "evaluator/data/*.yaml", "evaluator/data/**/*.yaml"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
